+++ src/spyglass/common/__init__.py
@@ -7,6 +7,7 @@ from spyglass.common.common_behav import (
     StateScriptFile,
     VideoFile,
     convert_epoch_interval_name_to_position_interval_name,
+    get_position_interval_epoch,
 )
 from spyglass.common.common_device import (
     CameraDevice,
@@ -40,10 +41,7 @@ from spyglass.common.common_interval import (
     intervals_by_length,
 )
 from spyglass.common.common_lab import Institution, Lab, LabMember, LabTeam
-from spyglass.common.common_nwbfile import (
-    AnalysisNwbfile,
-    Nwbfile,
-)
+from spyglass.common.common_nwbfile import AnalysisNwbfile, Nwbfile
 from spyglass.common.common_position import (
     IntervalLinearizationSelection,
     IntervalLinearizedPosition,